Move crypto/sshutil to an internal folder
