+++ internal/sshutil/inspect.go
